+++ .gitignore
@@ -9,3 +9,15 @@ Cargo.lock
 # These are backup files generated by rustfmt
 **/*.rs.bk
 .DS_Store
+
+# Environment variables
+.env
+
+# Editor
+.idea/
+.vscode/
+!.vscode/settings.json
+!.vscode/extensions.json
+
+# OpenAPI
+.openapi/
