+++ Cargo.toml
@@ -5,15 +5,15 @@ authors = ["Telcoin Team <engineering-backend@telco.in>"]
 edition = "2024"
 
 [dependencies]
-bon = "3.8"
-reqwest = { version = "0.13", features = ["json", "form"] }
-serde = { version = "1.0", features = ["derive"] }
-thiserror = "2.0"
-secrecy = "0.10"
 bigdecimal = { version = "0.4", features = ["serde"] }
+bon = "3"
+reqwest = { version = "0.13", features = ["json", "form"] }
+secrecy = "0.10"
+serde = { version = "1", features = ["derive"] }
+thiserror = "2"
 
 [dev-dependencies]
 dotenvy = "0.15"
+once_cell = "1"
 rand = "0.9"
-tokio = { version = "1.49", features = ["macros", "rt"] }
-once_cell = "1.21"
+tokio = { version = "1", features = ["macros", "rt"] }
