+++ src/models/mod.rs
@@ -1,17 +1,13 @@
-#![allow(missing_docs)]
-
-use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
-pub mod currency;
-pub mod links;
-pub mod requests;
-pub mod responses;
+use bon::Builder;
+use serde::{Deserialize, Serialize};
 
-pub use currency::*;
-use links::Links;
-pub use requests::*;
-pub use responses::*;
+mod currency;
+mod links;
+
+pub use currency::{Amount, Currency};
+pub use links::*;
 
 /// The details of a payment
 #[derive(Deserialize, Debug, Clone)]
@@ -107,6 +103,7 @@ pub struct PaymentDetails {
 /// See [PCI Compliance](https://docs.checkout.com/risk-management/pci-compliance)
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[serde(tag = "type")]
+#[allow(clippy::large_enum_variant)]
 pub enum PaymentRequestSource {
     /// A debit/credit/etc card
     #[serde(rename = "card")]
@@ -121,22 +118,27 @@ pub enum PaymentRequestSource {
         expiry_year: u32,
 
         /// The name of the cardholder
+        #[serde(skip_serializing_if = "Option::is_none")]
         name: Option<String>,
 
         /// The card verification value/code. 3 digits, except for Amex (4
         /// digits)
+        #[serde(skip_serializing_if = "Option::is_none")]
         cvv: Option<String>,
 
         /// This must be set to true for payments that use stored card details
         /// (default: false)
         ///
         /// See: [Requirements for stored payment details](https://docs.checkout.com/payments/store-payment-details/requirements-for-stored-payment-details)
+        #[serde(skip_serializing_if = "Option::is_none")]
         stored: Option<bool>,
 
         /// The billing address of the cardholder
+        #[serde(skip_serializing_if = "Option::is_none")]
         billing_address: Option<Address>,
 
         /// The phone number of the cardholder
+        #[serde(skip_serializing_if = "Option::is_none")]
         phone: Option<PhoneNumber>,
     },
 
@@ -158,6 +160,7 @@ pub enum PaymentRequestSource {
 /// The payout destination type
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[serde(tag = "type")]
+#[allow(clippy::large_enum_variant)]
 pub enum PaymentRequestDestination {
     /// A debit/credit/etc card
     #[serde(rename = "card")]
@@ -201,10 +204,12 @@ pub enum PaymentType {
     Moto,
 }
 
+/// Information about the sender of the payment's funds
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[serde(tag = "type")]
 pub enum PaymentSenderDetails {
     #[serde(rename = "individual")]
+    /// The individual sender
     Individual {
         /// The account holder's first name.
         ///
@@ -216,6 +221,7 @@ pub enum PaymentSenderDetails {
         first_name: String,
 
         /// The account holder's middle name.
+        #[serde(skip_serializing_if = "Option::is_none")]
         middle_name: Option<String>,
 
         /// The account holder's last name.
@@ -230,12 +236,15 @@ pub enum PaymentSenderDetails {
         /// The sender's date of birth, in the format YYYY-MM-DD.
         ///
         /// This field is required for cross-border money transfers.
+        #[serde(skip_serializing_if = "Option::is_none")]
         date_of_birth: Option<String>,
 
         /// The sender's country of birth, as a two-letter ISO country code.
+        #[serde(skip_serializing_if = "Option::is_none")]
         country_of_birth: Option<String>,
 
         /// The sender's nationality, as a two-letter ISO country code.
+        #[serde(skip_serializing_if = "Option::is_none")]
         nationality: Option<String>,
 
         /// The sender's registered corporate address.
@@ -253,6 +262,7 @@ pub enum PaymentSenderDetails {
     },
 
     #[serde(rename = "corporate")]
+    /// The corporate sender
     Corporate {
         /// The corporate sender's company name.
         ///
@@ -278,6 +288,7 @@ pub enum PaymentSenderDetails {
     },
 
     #[serde(rename = "government")]
+    /// The government sender
     Government {
         /// The sender's company name.
         ///
@@ -307,6 +318,7 @@ pub enum PaymentSenderDetails {
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[serde(tag = "type")]
 pub enum DestinationAccountHolder {
+    /// The individual account holder
     #[serde(rename = "individual")]
     Individual {
         /// The account holder's first name.
@@ -319,6 +331,7 @@ pub enum DestinationAccountHolder {
         ///
         /// This field is not required if the first name is already stored with
         /// the token or payment instrument id provided in destination.
+        #[serde(skip_serializing_if = "Option::is_none")]
         first_name: Option<String>,
 
         /// The account holder's last name.
@@ -331,12 +344,15 @@ pub enum DestinationAccountHolder {
         ///
         /// This field is not required if the last name is already stored with
         /// the token or payment instrument id provided in destination.
+        #[serde(skip_serializing_if = "Option::is_none")]
         last_name: Option<String>,
 
         /// The account holder's middle name.
+        #[serde(skip_serializing_if = "Option::is_none")]
         middle_name: Option<String>,
     },
 
+    /// The corporate account holder
     #[serde(rename = "corporate")]
     Corporate {
         /// The corporate account holder's company name.
@@ -349,6 +365,7 @@ pub enum DestinationAccountHolder {
         company_name: String,
     },
 
+    /// The government account holder
     #[serde(rename = "government")]
     Government {
         /// The government account holder's company name.
@@ -363,13 +380,15 @@ pub enum DestinationAccountHolder {
 }
 
 /// Additional details about the payout instruction.
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug, Clone, Builder)]
 pub struct DestinationInstruction {
     /// The funds transfer type code for the type of payout you're performing.
     ///
     /// You can only use codes that have been assigned to you by the card schemes and Checkout.com, based on your business case and requirements.
     ///
     /// If you have only been assigned one funds transfer type code, this field is optional.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub funds_transfer_type: Option<String>,
 
     /// The purpose of the payout.
@@ -387,12 +406,14 @@ pub struct DestinationInstruction {
     /// "medical_treatment", "emergency_need", "leisure", "savings", "gifts",
     /// "donations", "financial_services", "it_services", "investment", "insurance",
     /// "loan_payment", "pension", "royalties", "other", "income"
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub purpose: Option<PayoutPurpose>,
 }
 
 /// The purpose of a payout
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[serde(rename_all = "snake_case")]
+#[expect(missing_docs)]
 pub enum PayoutPurpose {
     FamilySupport,
     Expatriation,
@@ -416,103 +437,150 @@ pub enum PayoutPurpose {
 }
 
 /// A phone number
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug, Clone, Builder)]
 pub struct PhoneNumber {
     /// The international country calling code (1-7 characters)
+    #[builder(into)]
     pub country_code: String,
 
     /// The phone number (6-25 characters)
+    #[builder(into)]
     pub number: String,
 }
 
 /// A physical address
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug, Clone, Builder)]
 pub struct Address {
     /// The first line of the address (<= 200 characters)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub address_line1: Option<String>,
 
     /// The second line of the address (<= 200 characters)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub address_line2: Option<String>,
 
     /// The address city (<= 50 characters)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub city: Option<String>,
 
     /// The address state (<= 50 characters)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub state: Option<String>,
 
     /// The address zip/postal code (<= 50 characters)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub zip: Option<String>,
 
     /// The two-letter ISO country code of the address (2 characters)
-    pub country: Option<String>,
+    #[builder(into)]
+    pub country: String,
 }
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+/// The details of the customer
+#[derive(Serialize, Deserialize, Debug, Clone, Builder)]
 pub struct CustomerDescriptor {
     /// The identifier of an existing customer. (format: `cus_*`)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub id: Option<String>,
 
     /// The customer's email address. Providing this will create a new
     /// customer, unless you have already stored a customer with the same
     /// email.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub email: Option<String>,
 
     /// The customer's name. This will only set the name for new customers
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub name: Option<String>,
 }
 
 /// A description of the billing as it would appear on the account owner's
 /// statement
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug, Clone, Builder)]
 pub struct BillingDescriptor {
     /// A dynamic description of the charge (<= 25 characters)
+    #[builder(into)]
     pub name: String,
 
     /// The city from which the charge originated (1-13 characters)
+    #[builder(into)]
     pub city: String,
 }
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+/// The billing details
+#[derive(Serialize, Deserialize, Debug, Clone, Builder)]
+pub struct BillingInformation {
+    /// The billing address
+    pub address: Address,
+
+    /// The phone number associated with the billing address
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub phone: Option<PhoneNumber>,
+}
+
+/// The shipping details
+#[derive(Serialize, Deserialize, Debug, Clone, Builder)]
 pub struct ShippingDescriptor {
     /// The shipping address
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub address: Option<Address>,
 
     /// The phone number associated with the shipping address
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub phone: Option<PhoneNumber>,
 }
 
 /// Information for 3D Secure payments
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug, Clone, Builder)]
 pub struct _3DSRequest {
     /// Whether to process this payment as a 3D Secure payment (default: false)
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub enabled: Option<bool>,
 
     /// Determines whether to attempt a 3D Secure payment as non-3D Secure
     /// should the card issuer not be enrolled (default: false)
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub attempt_n3d: Option<bool>,
 
     /// The Electronic Commerce Indicator security level associated with the 3D
     /// Secure enrollment result. Required if using a third-party merchant
     /// plug-in (MPI) (<= 2 characters)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub sci: Option<String>,
 
     /// A Base64 encoded cryptographic identifier (CAVV) used by the card
     /// schemes to validate the cardholder authentication result (3D Secure).
     /// Required if using an external MPI (<= 50 characters)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub cryptogram: Option<String>,
 
     /// The 3D Secure transaction identifier. Required if using an external MPI
     /// with 3D Secure 2.X.X and a Mastercard card, or with 3D Secure 1.X.X for
     /// any supported card scheme (<= 100 characters)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub xid: Option<String>,
 
     /// Indicates the version of 3D Secure that was used for authentication.
     /// Defaults to 1.0.0 if not provided (<= 10 characters)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub version: Option<String>,
 
     /// Specifies an exemption reason so that the payment is not processed
     /// using 3D Secure authentication. Learn more about exemptions in our SCA
     /// compliance guide.
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub exemption: Option<ScaExemption>,
 }
 
@@ -542,33 +610,49 @@ pub enum ScaExemption {
     TrustedListing,
 }
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+/// Configures the risk assessment performed during the processing of the payment
+#[derive(Serialize, Deserialize, Debug, Clone, Builder)]
 pub struct RiskRequest {
     /// Whether a risk assessment should be performed (default: true)
     pub enabled: bool,
 }
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+/// Information about the recipient of the payment's funds
+///
+/// Relevant for [Account Funding Transactions](https://docs.checkout.com/payments/manage-payments/account-funding-transactions)
+/// and VISA or Mastercard [domestic UK transactions processed by financial institutions](https://docs.checkout.com/risk-management/requirements-for-financial-institutions).
+#[derive(Serialize, Deserialize, Debug, Clone, Builder)]
 pub struct PaymentRecipient {
     /// The recipient's date of birth (yyyy-mm-dd) (<= 10 characters)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub dob: Option<String>,
 
     /// The first six digits and the last four digits of the primary
     /// recipient's card, without spaces, or up to ten digits of the primary
     /// recipient's account number (10 characters)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub account_number: Option<String>,
 
     /// The first part of the UK postcode (e.g., W1T 4TJ would be W1T) (<= 50
     /// characters)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub zip: Option<String>,
 
     /// The recipient's first name (<= 50 characters)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub first_name: Option<String>,
 
     /// The recipient's last name (<= 50 characters)
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(into)]
     pub last_name: Option<String>,
 }
 
+/// Information for processing the payment
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct PaymentProcessingDescriptor {
     /// Indicates whether the payment is an Account Funding Transaction
@@ -610,56 +694,66 @@ pub struct PaymentProcessed {
     pub currency: Currency,
 
     /// Whether or not the authorization or capture was successful
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub approved: Option<bool>,
 
     /// The status of the payment
     pub status: PaymentStatus,
 
     /// The acquirer authorization code if the payment was authorized
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub auth_code: Option<String>,
 
     /// The Gateway response code
     pub response_code: String,
 
     /// The Gateway response summary
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub response_summary: Option<String>,
 
     /// Provides 3D Secure enrollment status if the payment was downgraded to
     /// non-3D Secure
-    #[serde(rename = "3ds")]
+    #[serde(rename = "3ds", skip_serializing_if = "Option::is_none")]
     pub three_ds: Option<_3dsStatus>,
 
     /// Returns the payment's risk assessment results
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub risk: Option<RiskResults>,
 
     /// The source of the payment
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub source: Option<PaymentProcessedSource>,
 
     /// The customer associated with the payment, if provided in the request
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub customer: Option<CustomerInfo>,
 
     /// The date/time the payment was processed
     pub processed_on: String,
 
     /// Your reference for the payment
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub reference: Option<String>,
 
     /// Returns information related to the processing of the payment
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub processing: Option<PaymentProcessingInfo>,
 
     /// The final Electronic Commerce Indicator (ECI) security level used to
     /// authorize the payment. Applicable for 3D Secure, digital wallet, and
     /// network token payments
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub eci: Option<String>,
 
     /// The scheme transaction identifier
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub scheme_id: Option<String>,
 
     /// The links related to the payment
     ///
     /// - Required: "self", "actions"
     /// - Optional: "void", "capture", "refund"
-    #[serde(rename = "_links")]
+    #[serde(rename = "_links", skip_serializing_if = "Option::is_none")]
     pub links: Option<Links>,
 }
 
@@ -674,21 +768,23 @@ pub struct PendingPayment {
     pub status: PaymentStatus,
 
     /// The customer associated with the payment, if provided in the request
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub customer: Option<CustomerInfo>,
 
     /// Your reference for the payment
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub reference: Option<String>,
 
     /// Provides 3D Secure enrollment status if the payment was downgraded to
     /// non-3D Secure
-    #[serde(rename = "3ds")]
+    #[serde(rename = "3ds", skip_serializing_if = "Option::is_none")]
     pub three_ds: Option<_3dsStatus>,
 
     /// The links related to the payment
     ///
     /// - Required: `"self"`
     /// - Optional: `"redirect"`
-    #[serde(rename = "_links")]
+    #[serde(rename = "_links", skip_serializing_if = "Option::is_none")]
     pub links: Option<Links>,
 }
 
@@ -696,6 +792,7 @@ pub struct PendingPayment {
 ///
 /// See: [Get Payment Details](https://docs.checkout.com/payments/manage-payments/get-payment-details)
 #[derive(Serialize, Deserialize, Debug, Copy, Clone, PartialEq, Eq, Default)]
+#[expect(missing_docs)]
 pub enum PaymentStatus {
     Authorized,
     #[default]
@@ -728,23 +825,29 @@ pub struct _3dsStatus {
     /// Verification to ensure the integrity of the response
     ///
     /// Example: `"Y"`
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub signature_valid: Option<String>,
 
     /// Indicates whether or not the cardholder was authenticated
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub authentication_response: Option<_3dsAuthenticationStatus>,
 
     /// Base64 encoded cryptographic identifier (CAVV) used by the card schemes
     /// to validate the integrity of the 3D secure payment data
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub cryptogram: Option<String>,
 
     /// Unique identifier for the transaction assigned by the MPI
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub xid: Option<String>,
 
     /// Indicates the version of 3D Secure that was used for authentication
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub version: Option<String>,
 
     /// Specifies an exemption reason so that the payment is not processed
     /// using 3D Secure authentication
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub exemption: Option<ScaExemption>,
 }
 
@@ -798,6 +901,7 @@ pub struct RiskResults {
 /// will be the name of the alternative payment method
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[serde(tag = "type")]
+#[allow(clippy::large_enum_variant)]
 pub enum PaymentProcessedSource {
     /// A debit/credit/etc card
     #[serde(rename = "card")]
@@ -805,12 +909,15 @@ pub enum PaymentProcessedSource {
         /// The payment source identifier that can be used for subsequent
         /// payments. For new sources, this will only be returned if the
         /// payment was approved
+        #[serde(skip_serializing_if = "Option::is_none")]
         id: Option<String>,
 
         /// The payment source owner's billing address
+        #[serde(skip_serializing_if = "Option::is_none")]
         billing_address: Option<Address>,
 
         /// The payment source owner's phone number
+        #[serde(skip_serializing_if = "Option::is_none")]
         phone: Option<PhoneNumber>,
 
         /// The expiry month (1-2 characters)
@@ -820,9 +927,11 @@ pub enum PaymentProcessedSource {
         expiry_year: u32,
 
         /// The cardholder's name
+        #[serde(skip_serializing_if = "Option::is_none")]
         name: Option<String>,
 
         /// The card scheme
+        #[serde(skip_serializing_if = "Option::is_none")]
         scheme: Option<String>,
 
         /// The last four digits of the card number
@@ -837,36 +946,46 @@ pub enum PaymentProcessedSource {
         bin: String,
 
         /// The card type
+        #[serde(skip_serializing_if = "Option::is_none")]
         card_type: Option<CardType>,
 
         /// The card category
+        #[serde(skip_serializing_if = "Option::is_none")]
         card_category: Option<CardCategory>,
 
         /// The name of the card issuer
+        #[serde(skip_serializing_if = "Option::is_none")]
         issuer: Option<String>,
 
         /// The card issuer's country (two-letter ISO code) (2 characters)
+        #[serde(skip_serializing_if = "Option::is_none")]
         issuer_country: Option<String>,
 
         /// The issuer/card scheme product identifier
+        #[serde(skip_serializing_if = "Option::is_none")]
         product_id: Option<String>,
 
         /// The issuer/card scheme product type
+        #[serde(skip_serializing_if = "Option::is_none")]
         product_type: Option<String>,
 
         /// The card verification value (CVV) check result
+        #[serde(skip_serializing_if = "Option::is_none")]
         cvv_result: Option<String>,
 
         /// Whether the card supports payouts
+        #[serde(skip_serializing_if = "Option::is_none")]
         payouts: Option<bool>,
 
         /// The fast funds eligibility of the card
         ///
         /// See: [Card Payouts](https://docs.checkout.com/card-payouts)
+        #[serde(skip_serializing_if = "Option::is_none")]
         fast_funds: Option<bool>,
 
         /// A unique reference to the underlying card for network tokens (e.g.
         /// Apple Pay, Google Pay)
+        #[serde(skip_serializing_if = "Option::is_none")]
         payment_account_reference: Option<String>,
     },
 
@@ -882,6 +1001,7 @@ pub enum PaymentProcessedSource {
 /// The processed payment's destination type
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[serde(tag = "type")]
+#[allow(clippy::large_enum_variant)]
 pub enum PaymentProcessedDestination {
     /// A debit/credit/etc card
     #[serde(rename = "card")]
@@ -889,12 +1009,15 @@ pub enum PaymentProcessedDestination {
         /// The payment source identifier that can be used for subsequent
         /// payments. For new sources, this will only be returned if the
         /// payment was approved
+        #[serde(skip_serializing_if = "Option::is_none")]
         id: Option<String>,
 
         /// The payment source owner's billing address
+        #[serde(skip_serializing_if = "Option::is_none")]
         billing_address: Option<Address>,
 
         /// The payment source owner's phone number
+        #[serde(skip_serializing_if = "Option::is_none")]
         phone: Option<PhoneNumber>,
 
         /// The expiry month (1-2 characters)
@@ -904,9 +1027,11 @@ pub enum PaymentProcessedDestination {
         expiry_year: u32,
 
         /// The cardholder's name
+        #[serde(skip_serializing_if = "Option::is_none")]
         name: Option<String>,
 
         /// The card scheme
+        #[serde(skip_serializing_if = "Option::is_none")]
         scheme: Option<String>,
 
         /// The last four digits of the card number
@@ -921,36 +1046,46 @@ pub enum PaymentProcessedDestination {
         bin: String,
 
         /// The card type
+        #[serde(skip_serializing_if = "Option::is_none")]
         card_type: Option<CardType>,
 
         /// The card category
+        #[serde(skip_serializing_if = "Option::is_none")]
         card_category: Option<CardCategory>,
 
         /// The name of the card issuer
+        #[serde(skip_serializing_if = "Option::is_none")]
         issuer: Option<String>,
 
         /// The card issuer's country (two-letter ISO code) (2 characters)
+        #[serde(skip_serializing_if = "Option::is_none")]
         issuer_country: Option<String>,
 
         /// The issuer/card scheme product identifier
+        #[serde(skip_serializing_if = "Option::is_none")]
         product_id: Option<String>,
 
         /// The issuer/card scheme product type
+        #[serde(skip_serializing_if = "Option::is_none")]
         product_type: Option<String>,
 
         /// The card verification value (CVV) check result
+        #[serde(skip_serializing_if = "Option::is_none")]
         cvv_result: Option<String>,
 
         /// Whether the card supports payouts
+        #[serde(skip_serializing_if = "Option::is_none")]
         payouts: Option<bool>,
 
         /// The fast funds eligibility of the card
         ///
         /// See: [Card Payouts](https://docs.checkout.com/card-payouts)
+        #[serde(skip_serializing_if = "Option::is_none")]
         fast_funds: Option<bool>,
 
         /// A unique reference to the underlying card for network tokens (e.g.
         /// Apple Pay, Google Pay)
+        #[serde(skip_serializing_if = "Option::is_none")]
         payment_account_reference: Option<String>,
     },
 }
@@ -959,11 +1094,16 @@ pub enum PaymentProcessedDestination {
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[serde(rename_all = "UPPERCASE")]
 pub enum CardType {
+    /// Credit card
     Credit,
+    /// Debit card
     Debit,
+    /// Prepaid card
     Prepaid,
+    /// Charge card
     Charge,
     #[serde(rename = "DEFERRED DEBIT")]
+    /// Deferred debit card
     DeferredDebit,
 }
 
@@ -971,7 +1111,9 @@ pub enum CardType {
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[serde(rename_all = "UPPERCASE")]
 pub enum CardCategory {
+    /// Consumer card
     Consumer,
+    /// Commercial card
     Commercial,
 }
 
@@ -983,19 +1125,24 @@ pub struct CustomerInfo {
     pub id: String,
 
     /// The customer's email address
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub email: Option<String>,
 
     /// The customer's name
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub name: Option<String>,
 }
 
+/// Returns information related to the processing of the payment
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct PaymentProcessingInfo {
     /// A unique identifier for the authorization that is submitted to the card
     /// scheme during processing
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub retrieval_reference_number: Option<String>,
 
     /// A unique identifier for the transaction generated by the acquirer
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub acquirer_transaction_id: Option<String>,
 }
 
@@ -1013,9 +1160,11 @@ pub struct ActionSummary {
     response_code: String,
 
     /// The Gateway response summary
+    #[serde(skip_serializing_if = "Option::is_none")]
     response_summary: Option<String>,
 }
 
+/// The full details of a payment action
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct Action {
     /// The unique identifier of the payment action (format: `act_*`)
@@ -1032,43 +1181,53 @@ pub struct Action {
     amount: u64,
 
     /// Whether the action was successful
+    #[serde(skip_serializing_if = "Option::is_none")]
     approved: Option<bool>,
 
     /// The acquirer authorization code for cards
+    #[serde(skip_serializing_if = "Option::is_none")]
     auth_code: Option<String>,
 
     /// The Gateway response code
     response_code: String,
 
     /// The Gateway response summary
+    #[serde(skip_serializing_if = "Option::is_none")]
     response_summary: Option<String>,
 
     /// Your reference for the action
+    #[serde(skip_serializing_if = "Option::is_none")]
     reference: Option<String>,
 
     /// Returns information related to the processing of the payment
+    #[serde(skip_serializing_if = "Option::is_none")]
     processing: Option<ActionProcessingInfo>,
 
     /// A set of key-value pairs that you can attach to an action
     metadata: Metadata,
 }
 
+/// Returns information related to the processing of the action
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct ActionProcessingInfo {
     /// A unique identifier for the authorization that is submitted to the card
     /// scheme during processing
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub retrieval_reference_number: Option<String>,
 
     /// A unique identifier for the capture that is submitted to the card
     /// scheme during processing
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub acquirer_reference_number: Option<String>,
 
     /// A unique identifier for the transaction generated by the acquirer
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub acquirer_transaction_id: Option<String>,
 }
 
 /// The type of an action
 #[derive(Serialize, Deserialize, Debug, Clone)]
+#[expect(missing_docs)]
 pub enum ActionType {
     Authorization,
     #[serde(rename = "Card Verification")]
@@ -1082,6 +1241,7 @@ pub enum ActionType {
 /// The source to get card metadata
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[serde(tag = "type")]
+#[expect(missing_docs)]
 pub enum CardMetadataSource {
     #[serde(rename = "card")]
     Card {
@@ -1110,6 +1270,9 @@ pub enum CardMetadataSource {
     },
 }
 
+/// Returns a single metadata record for the card specified by the Primary
+/// Account Number (PAN), Bank Identification Number (BIN), token, or instrument
+/// supplied.
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub enum CardMetadataFormat {
     /// A basic response will only include standard metadata
@@ -1120,16 +1283,4 @@ pub enum CardMetadataFormat {
     /// card payouts.
     #[serde(rename = "card_payouts")]
     CardPayouts,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct CardMetadataRequest {
-    /// The source object
-    pub source: CardMetadataSource,
-
-    /// The format to provide the output in.
-    ///
-    /// Default is "basic"
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub format: Option<CardMetadataFormat>,
 }
